# improve(smt/pkg/utils): improve the performance of HashContractBytecodeBigInt()