Changes:
1. replace [vectorized-poseidon-gold](https://github.com/gateway-fm/vectorized-poseidon-gold) by OKX [poseidongold](https://github.com/okx/poseidongold) (this leads to around 25% improvement). poseidongold repo implements a Go wrapper of Rust-based Poseidon hashing from Plonky2. This is 2X faster than vectorized-poseidon-gold from gateway-fm.
2. remove string concatenations and implement custom conversion in HashContractBytecodeBigInt() (this leads to around 2X improvement). The combined effect of these two changes is 2X speedup for HashContractBytecodeBigInt().

To test the performance, you can do:
```
$ cd smt/pkg/utils
$ go test -benchmem -run=^$ -bench ^BenchmarkHashContractBytecode$ .
```

For example:
```
goos: linux
goarch: amd64
pkg: github.com/ledgerwatch/erigon/smt/pkg/utils
cpu: AMD Ryzen 9 7950X 16-Core Processor            
BenchmarkHashContractBytecode/1-32         38618             28042 ns/op           15072 B/op        563 allocs/op
BenchmarkHashContractBytecode/2-32        119170             10211 ns/op            3528 B/op         53 allocs/op
BenchmarkHashContractBytecode/3-32        110352             10487 ns/op            3736 B/op         56 allocs/op
PASS
```

To check the performance stated at item 1 above, you can do:
```
$ git clone https://github.com/okx/poseidongold.git
$ cd poseidongold/go
$ go test -benchmem -run=^$ -bench .
```

For example:
```
goos: linux
goarch: amd64
pkg: github.com/okx/poseidongold/go
cpu: AMD Ryzen 9 7950X 16-Core Processor            
BenchmarkHashWithResult-32               1314997               881.7 ns/op             0 B/op          0 allocs/op
BenchmarkVectorizedPoseidongold-32        763915              1530 ns/op               0 B/op          0 allocs/op
PASS
```